requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
selectolax==0.3.17
python-dotenv==1.0.0
cachetools==5.3.2
